Fixed an issue with changing the values of cells
When changing the value of a cell, it doesn't change the number of that cell. For example, if a cell is first a Start then a Wall; the start counter doesn't change and so another Start cannot be placed. This will be fixed by blocking the user if the cell already has a value.